Moved utility to seperate folder
