+++ include/BSMPT/baryo_calculation/Fluid_Type/gen_func_fluid.h
@@ -1,5 +1,6 @@
 // Copyright (C) 2020  Philipp Basler, Margarete Mühlleitner and Jonas Müller
-// SPDX-FileCopyrightText: 2021 Philipp Basler, Margarete Mühlleitner and Jonas Müller
+// SPDX-FileCopyrightText: 2021 Philipp Basler, Margarete Mühlleitner and Jonas
+// Müller
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
@@ -11,7 +12,7 @@
 #include <BSMPT/baryo_calculation/transport_equations.h>
 #include <BSMPT/models/IncludeAllModels.h>
 #include <BSMPT/models/SMparam.h>
-#include <BSMPT/utility.h>
+#include <BSMPT/utility/utility.h>
 #include <boost/numeric/odeint.hpp>
 #include <iostream>
