+++ src/models/ModelTestfunctions.cpp
@@ -5,7 +5,7 @@
 
 #include <BSMPT/models/ClassPotentialOrigin.h>
 #include <BSMPT/models/ModelTestfunctions.h>
-#include <BSMPT/utility.h>
+#include <BSMPT/utility/utility.h>
 #include <iomanip>
 
 namespace BSMPT
