+++ tests/Test-r2hdm.cpp
@@ -9,7 +9,6 @@
 #include <BSMPT/models/ClassPotentialOrigin.h> // for Class_Potential_Origin
 #include <BSMPT/models/IncludeAllModels.h>
 #include <BSMPT/models/ModelTestfunctions.h>
-#include <BSMPT/utility.h>
 
 #include "GenerateTestCompares/R2HDM.h"
