+++ tests/Test-rn2hdm.cpp
@@ -9,7 +9,7 @@
 #include <BSMPT/models/ClassPotentialOrigin.h> // for Class_Potential_Origin
 #include <BSMPT/models/IncludeAllModels.h>
 #include <BSMPT/models/ModelTestfunctions.h>
-#include <BSMPT/utility.h>
+#include <BSMPT/utility/utility.h>
 
 #include "GenerateTestCompares/RN2HDM.h"
